feat: add error handler in controllers

+++ src/controllers/instructor.ts
@@ -1,17 +1,52 @@
 import { Request, Response, NextFunction } from "express";
 import { InstrutorModel } from "../models/instrutor";
 import { StatusCodes } from "http-status-codes";
+import mongoose from "mongoose";
 
 // Gets all instructors data
-export const getAllInstructors = async (req: Request, res: Response) => {
+export const getAllInstructors = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
   const Instructors = await InstrutorModel.find({});
+  if (!Instructors) {
+    return next(
+      res.status(StatusCodes.NOT_FOUND).json({ msg: "No instructors found." })
+    );
+  }
+
   res.status(StatusCodes.OK).json({ Instructors });
 };
 
 // Creates new instructor
-export const createInstructor = async (req: Request, res: Response) => {
+export const createInstructor = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  const { nome, email, password, especialidades, horariosDisponiveis } =
+    req.body;
+
+  if (!nome || !email || !password || !especialidades || !horariosDisponiveis) {
+    return next(
+      res.status(StatusCodes.BAD_REQUEST).json({ msg: "Missing fields" })
+    );
+  }
+
+  const emailAlreadyExists = await InstrutorModel.findOne({ email });
+  if (emailAlreadyExists) {
+    return next(
+      res.status(StatusCodes.BAD_REQUEST).json({ msg: "Email already exists" })
+    );
+  }
+
   const Instructor = await InstrutorModel.create({
-    ...req.body,
+    nome,
+    email,
+    password,
+    especialidades,
+    horariosDisponiveis,
     role: "Instructor",
   });
   res.status(StatusCodes.CREATED).json({ Instructor });
@@ -23,25 +58,42 @@ export const updateInstructor = async (
   res: Response,
   next: NextFunction
 ) => {
-  const { id: InstrutorId } = req.params;
-  const Instrutor = await InstrutorModel.findOneAndUpdate(
-    {
-      _id: InstrutorId,
-    },
-    req.body,
-    {
-      new: true,
-      runValidators: true,
+  try {
+    const { id: InstrutorId } = req.params;
+
+    const InstructorDocument = await InstrutorModel.findById(InstrutorId);
+    if (!InstructorDocument) {
+      return next(
+        res
+          .status(StatusCodes.NOT_FOUND)
+          .json({ msg: `there is no instructor with id: ${InstrutorId}` })
+      );
     }
-  );
-  if (!Instrutor) {
-    return next(
-      res
-        .status(StatusCodes.NOT_FOUND)
-        .json({ msg: `There is no instructor with id: ${InstrutorId}` })
+
+    const instructor = await InstrutorModel.findOneAndUpdate(
+      { _id: InstrutorId },
+      {
+        ...req.body,
+      }
     );
+    if (!instructor) {
+      return next(
+        res
+          .status(StatusCodes.BAD_REQUEST)
+          .json({ msg: "Failed to update instructor, please try again." })
+      );
+    }
+
+    res.status(StatusCodes.OK).json({ instructor });
+  } catch (error) {
+    if (error instanceof mongoose.Error.CastError) {
+      return next(
+        res.status(StatusCodes.BAD_REQUEST).json({
+          msg: "Please inform a valid id.",
+        })
+      );
+    }
   }
-  res.status(StatusCodes.OK).json({ Instrutor });
 };
 
 // Deletes instructor data
@@ -50,14 +102,34 @@ export const deleteInstructor = async (
   res: Response,
   next: NextFunction
 ) => {
-  const { id: InstrutorId } = req.params;
-  const Instructor = await InstrutorModel.findByIdAndRemove(InstrutorId);
-  if (!Instructor) {
-    return next(
-      res
-        .status(StatusCodes.NOT_FOUND)
-        .json({ msg: `There is no instructor with id: ${InstrutorId}` })
-    );
+  try {
+    const { id: InstrutorId } = req.params;
+
+    const InstructorDocument = await InstrutorModel.findById(InstrutorId);
+    if (!InstructorDocument) {
+      return next(
+        res
+          .status(StatusCodes.NOT_FOUND)
+          .json({ msg: `there is no instructor with id: ${InstrutorId}` })
+      );
+    }
+
+    const instructor = await InstrutorModel.findByIdAndRemove(InstrutorId);
+    if (!instructor) {
+      return next(
+        res
+          .status(StatusCodes.BAD_REQUEST)
+          .json({ msg: "Failed to delete instructor. Please try again." })
+      );
+    }
+    res.status(StatusCodes.OK).json("Deleted!");
+  } catch (error) {
+    if (error instanceof mongoose.Error.CastError) {
+      return next(
+        res.status(StatusCodes.BAD_REQUEST).json({
+          msg: "Please inform a valid id.",
+        })
+      );
+    }
   }
-  res.status(StatusCodes.OK).json("Deleted!");
 };
